requirements.txt
# ── Core Vanna 2.0 with SQLite ────────────────────────────────────────────────
vanna[sqlite]>=2.0.0

# ── Web framework ─────────────────────────────────────────────────────────────
fastapi
uvicorn[standard]

# ── Data & visualisation ──────────────────────────────────────────────────────
plotly
pandas

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv

# ── Rate limiting (bonus) ─────────────────────────────────────────────────────
slowapi
